pyproject.toml
[tool.poetry]
name = "tap-clickhouse"
version = "0.0.1"
description = "`tap-clickhouse` is a Singer tap for ClickHouse, built with the Meltano Singer SDK."
readme = "README.md"
authors = ["Alik Kurdyukov <alik@kurdyukov.com>"]
keywords = [
    "ELT",
    "ClickHouse",
]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = ">=3.7.1,<4"
singer-sdk = { version="~=0.32.0" }
fs-s3fs = { version = "~=1.1.1", optional = true }
clickhouse_sqlalchemy = { git = "https://github.com/akurdyukov/clickhouse-sqlalchemy.git" }

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.0"
singer-sdk = { version="~=0.32.0", extras = ["testing"] }

[tool.poetry.extras]
s3 = ["fs-s3fs"]

[tool.mypy]
python_version = "3.9"
warn_unused_configs = true
plugins = "sqlmypy"

[tool.ruff]
ignore = [
    "ANN101",  # missing-type-self
    "ANN102",  # missing-type-cls
]
select = ["ALL"]
src = ["tap_clickhouse"]
target-version = "py37"


[tool.ruff.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.isort]
known-first-party = ["tap_clickhouse"]

[tool.ruff.pydocstyle]
convention = "google"

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
# CLI declaration
tap-clickhouse = 'tap_clickhouse.tap:TapClickHouse.cli'
